cross entropy

+++ train.py
@@ -48,7 +48,7 @@ def train():
         # forward + backward + optimize
         outputs = net(inputs)
         torch.where(outputs > 0.5, torch.ones(1).cuda(),torch.zeros(1).cuda())
-        loss = criterion(input=outputs,target=masks)
+        loss = criterion(input=outputs,target=masks.to(torch.long))
         loss.backward()
         optimizer.step()
